fix: Resolve CS1503 in GetScreenRefreshRate
- Fixed a CS1503 build error in OverlayService.cs by explicitly casting ENUM_CURRENT_SETTINGS from int to uint when calling the EnumDisplaySettingsEx P/Invoke function within GetScreenRefreshRate.

+++ ConditioningControlPanel/Services/OverlayService.cs
@@ -1021,7 +1021,7 @@ private BitmapSource? CaptureScreen(System.Windows.Forms.Screen screen)
     {
         DEVMODE dm = new DEVMODE();
         dm.dmSize = (ushort)System.Runtime.InteropServices.Marshal.SizeOf(typeof(DEVMODE));
-        if (EnumDisplaySettingsEx(screen.DeviceName, ENUM_CURRENT_SETTINGS, ref dm, 0))
+        if (EnumDisplaySettingsEx(screen.DeviceName, (uint)ENUM_CURRENT_SETTINGS, ref dm, 0))
         {
             return (int)dm.dmDisplayFrequency;
         }
